requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-redis==5.4.0
django-prometheus==2.3.1
django-rq==3.0.1
drf-spectacular==0.29.0
python-dotenv==1.2.1
colorlog==6.10.1
python-json-logger==2.0.7
PyJWT==2.11.0
redis==6.4.0
reportlab==4.4.10
pypdf==6.7.2
pillow==12.1.1
